changed name of msghub producer cmd
